refactor mdx articles types
